state: delete Slice<T>, move persistence under observable
After the three caller migrations, Slice<T> has no users left. This
commit:

- Deletes Slice, SliceWriteGuard, and the SliceEventEmitter trait from
  state/slice/mod.rs. The three slice-level tests are absorbed by
  observable's own tests; persistence + registry tests move with their
  respective code.
- Moves apps/native/src-tauri/src/observable.rs to
  src/observable/mod.rs and pulls persistence.rs + json_io.rs under it.
  The Persistence trait + AppDataJson + RepoScopedJson are re-exported
  at crate::observable::*, so callers say
  `use crate::observable::{AppDataJson, Persistence, Observable}`.
- Shrinks state/slice/mod.rs to a thin re-export of registry.rs. The
  module stays as the home for SliceRegistry until B1 retires it.
- Updates state/mod.rs prose to describe the new shape (observables, not
  slices).
