+++ apps/native/src-tauri/src/state/slice/registry.rs
@@ -88,3 +88,40 @@ impl SliceRegistry {
         (entry.set_field_fn)(app, field_key, value)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn registry_exposes_registered_slice_configs() {
+        fn schema_stub(
+            _: &tauri::AppHandle<tauri::Wry>,
+        ) -> Result<configurable::ConfigurableSchema> {
+            unreachable!("schema is not invoked by this registry test")
+        }
+
+        fn set_stub(_: &tauri::AppHandle<tauri::Wry>, _: &str, _: serde_json::Value) -> Result<()> {
+            unreachable!("set is not invoked by this registry test")
+        }
+
+        let registry = SliceRegistry::default();
+        registry
+            .register(RegisteredSliceConfig {
+                name: "DemoState",
+                schema_fn: schema_stub,
+                set_field_fn: set_stub,
+            })
+            .expect("slice config registers");
+
+        let entries = registry.entries().expect("registry entries load");
+        assert_eq!(entries.len(), 1);
+        assert_eq!(entries[0].name, "DemoState");
+        let _schema_fn = entries[0].schema_fn;
+        let _set_field_fn = entries[0].set_field_fn;
+        assert!(registry
+            .get("DemoState")
+            .expect("registry lookup")
+            .is_some());
+    }
+}
